+++ package.json
@@ -10,8 +10,14 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.6",
     "react": "16.9.0",
-    "react-native": "0.61.5"
+    "react-native": "0.61.5",
+    "react-native-gesture-handler": "^1.5.6",
+    "react-native-safe-area-context": "^0.6.4",
+    "react-native-screens": "^2.0.0-beta.1",
+    "react-navigation": "^4.1.1",
+    "react-navigation-stack": "^2.1.1"
   },
   "devDependencies": {
     "@babel/core": "7.8.3",
